Perform caching when accessing through origins
