+++ src/main/java/sushi/compile/path_condition_distance/SimilarityWithNumericExpression.java
@@ -20,7 +20,7 @@ public class SimilarityWithNumericExpression implements ClauseSimilarityHandler 
 	}
 
 	@Override
-	public double evaluateSimilarity(CandidateBackbone backbone, Map<String, Object> candidateObjects, Map<Long, String> constants) {
+	public double evaluateSimilarity(CandidateBackbone backbone, Map<String, Object> candidateObjects, Map<Long, String> constants, SushiLibCache cache) {
 		logger.debug("Handling similarity for numeric expression");
 		
 		double similarity = 0.0d;
@@ -29,7 +29,7 @@ public class SimilarityWithNumericExpression implements ClauseSimilarityHandler 
 			final ArrayList<Object> variables = new ArrayList<>();
 			for (String variableOrigin : this.theValueCalculator.getVariableOrigins()) {
 				theVariableOrigin = variableOrigin;
-				Object variableValue = backbone.retrieveOrVisitField(variableOrigin, candidateObjects, constants);
+				Object variableValue = backbone.retrieveOrVisitField(variableOrigin, candidateObjects, constants, cache);
 				variables.add(variableValue);
 			}
 			similarity += inverseDistanceRatio(this.theValueCalculator.calculate(variables), 1.0d);
